+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 0.0.2 - 28-10-2021
+  * add missing utils test
+  * update README
+  * add verify webhook signature api
 ## 0.0.1 - 28-10-2021
   - Initial Release